Moved generic Zotero methods into an abstract job class.
