+++ src/Job/Import.php
@@ -1,77 +1,8 @@
 <?php
 namespace ZoteroImport\Job;
 
-use DateTime;
-use Omeka\Job\AbstractJob;
-use Omeka\Job\Exception;
-use Zend\Http\Client;
-use Zend\Http\Response;
-use Zend\Log\Logger;
-use ZoteroImport\Zotero\Url;
-
-class Import extends AbstractJob
+class Import extends AbstractZoteroSync
 {
-    const ACTION_CREATE = 'create'; // @translate
-    const ACTION_REPLACE = 'replace'; // @translate
-
-    /**
-     * Number of resources to process by batch (Zotero default limit is 50).
-     *
-     * @var int
-     */
-    protected $sizeChunk = 50;
-
-    /**
-     * @var Logger
-     */
-    protected $logger;
-
-    /**
-     * Zotero API client
-     *
-     * @var Client
-     */
-    protected $client;
-
-    /**
-     * Zotero API URL
-     *
-     * @var Url
-     */
-    protected $url;
-
-    /**
-     * Vocabularies to cache.
-     *
-     * @var array
-     */
-    protected $vocabularies = [
-        'dcterms' => 'http://purl.org/dc/terms/',
-        'dctype'  => 'http://purl.org/dc/dcmitype/',
-        'bibo'    => 'http://purl.org/ontology/bibo/',
-    ];
-
-    /**
-     * Cache of selected Omeka resource classes
-     *
-     * @var array
-     */
-    protected $resourceClasses = [];
-
-    /**
-     * Cache of selected Omeka properties
-     *
-     * @var array
-     */
-    protected $properties = [];
-
-    /**
-     * Priority map between Zotero item types and Omeka resource classes
-     *
-     * @var array
-     */
-    protected $itemTypeMap = [];
-
     /**
      * Priority map between Zotero item fields and Omeka properties
      *
@@ -132,7 +63,6 @@ class Import extends AbstractJob
         $this->setImportUrl();
 
         $apiVersion = $this->getArg('version', 0);
-        $apiKey = $this->getArg('apiKey');
         $collectionKey = $this->getArg('collectionKey');
         $action = $this->getArg('action');
 
@@ -158,72 +88,14 @@ class Import extends AbstractJob
             return;
         }
 
-        // Cache all Zotero parent and child items.
-        $zParentItems = [];
-        $zChildItems = [];
-        foreach (array_chunk($zItemKeys, $this->sizeChunk, true) as $zItemKeysChunk) {
-            if ($this->shouldStop()) {
-                return;
-            }
-            $url = $this->url->items([
-                'itemKey' => implode(',', $zItemKeysChunk),
-                // Include the Zotero key so Zotero adds enclosure links to the
-                // response. An attachment can only be downloaded if an
-                // enclosure link is included.
-                'key' => $apiKey,
-            ]);
-            $zItems = json_decode($this->getResponse($url)->getBody(), true);
-
-            foreach ($zItems as $zItem) {
-                $dateAdded = new DateTime($zItem['data']['dateAdded']);
-                if ($dateAdded->getTimestamp() < $this->getArg('timestamp', 0)) {
-                    // Only import items added since the passed timestamp. Note
-                    // that the timezone must be UTC.
-                    continue;
-                }
-
-                // Unset unneeded data to save memory.
-                unset($zItem['library']);
-                unset($zItem['version']);
-                unset($zItem['meta']);
-                unset($zItem['links']['self']);
-                unset($zItem['links']['alternate']);
-
-                if (isset($zItem['data']['parentItem'])) {
-                    $zChildItems[$zItem['data']['parentItem']][] = $zItem;
-                } else {
-                    $zParentItems[$zItem['key']] = $zItem;
-                }
-            }
-        }
+        list($zParentItems, $zChildItems) = $this->fetchZoteroItems($zItemKeys);
 
         switch ($action) {
-            // Get the Omeka item ids of Zotero items that are already imported.
             case self::ACTION_REPLACE:
-                /** @var \Doctrine\DBAL\Connection $connection */
-                $connection = $services->get('Omeka\Connection');
-                $qb = $connection->createQueryBuilder();
-                // TODO How to do a "WHERE IN" with doctrine and strings?
-                $quotedZParentItemKeys = array_map([$connection, 'quote'], array_keys($zParentItems));
-                $qb
-                    ->select([
-                        // Should be the first column.
-                        'zotero_key' => 'zotero_import_item.zotero_key',
-                        'item_id' => 'zotero_import_item.item_id',
-                    ])
-                    ->from('zotero_import_item', 'zotero_import_item')
-                    // ->where($qb->expr()->in('zotero_import_item.zotero_key', ':zotero_keys'))
-                    // ->setParameter('zotero_keys', array_keys($zParentItems))
-                    ->where($qb->expr()->in('zotero_import_item.zotero_key', $quotedZParentItemKeys))
-                    // Only one identifier by resource, and the first one.
-                    ->groupBy(['zotero_import_item.zotero_key'])
-                    ->orderBy('zotero_import_item.id', 'ASC');
-                $stmt = $connection->executeQuery($qb, $qb->getParameters());
-                $existingItems = $stmt->fetchAll(\PDO::FETCH_KEY_PAIR);
+                $existingItems = $this->existingItems(array_keys($zParentItems));
                 // Keep order of the keys provided by Zotero.
                 // TODO Order by "date modified" in Zotero?
                 $existingItems = array_intersect_key(
-
                     array_replace($zParentItems, $existingItems),
                     $existingItems
                 );
@@ -315,107 +187,13 @@ class Import extends AbstractJob
     }
 
     /**
-     * Set the HTTP client to use during this import.
-     */
-    public function setImportClient()
-    {
-        $headers = ['Zotero-API-Version' => '3'];
-        if ($apiKey = $this->getArg('apiKey')) {
-            $headers['Authorization'] = sprintf('Bearer %s', $apiKey);
-        }
-        $this->client = $this->getServiceLocator()->get('Omeka\HttpClient')
-            ->setHeaders($headers)
-            // Decrease the chance of timeout by increasing to 20 seconds,
-            // which splits the time between Omeka's default (10) and Zotero's
-            // upper limit (30).
-            ->setOptions(['timeout' => 20]);
-    }
-
-    /**
-     * Set the Zotero URL object to use during this import.
-     */
-    public function setImportUrl()
-    {
-        $this->url = new Url($this->getArg('type'), $this->getArg('id'));
-    }
-
-    /**
-     * Get a response from the Zotero API.
-     *
-     * @param string $url
-     * @return Response
-     */
-    public function getResponse($url)
-    {
-        $response = $this->client->setUri($url)->send();
-        if (!$response->isSuccess()) {
-            throw new Exception\RuntimeException(sprintf(
-                'Requested "%s" got "%s".', $url, $response->renderStatusLine()
-            ));
-        }
-        return $response;
-    }
-
-    /**
-     * Cache selected resource classes.
-     */
-    public function cacheResourceClasses()
-    {
-        $api = $this->getServiceLocator()->get('Omeka\ApiManager');
-        foreach ($this->vocabularies as $prefix => $namespaceUri) {
-            $classes = $api->search('resource_classes', [
-                'vocabulary_namespace_uri' => $namespaceUri,
-            ])->getContent();
-            foreach ($classes as $class) {
-                $this->resourceClasses[$prefix][$class->localName()] = $class;
-            }
-        }
-    }
-
-    /**
-     * Cache selected properties.
-     */
-    public function cacheProperties()
-    {
-        $api = $this->getServiceLocator()->get('Omeka\ApiManager');
-        foreach ($this->vocabularies as $prefix => $namespaceUri) {
-            $properties = $api->search('properties', [
-                'vocabulary_namespace_uri' => $namespaceUri,
-            ])->getContent();
-            foreach ($properties as $property) {
-                $this->properties[$prefix][$property->localName()] = $property;
-            }
-        }
-    }
-
-    /**
-     * Convert a mapping with terms into a mapping with prefix and local name.
-     *
-     * @param string $mapping
-     * @return array
-     */
-    protected function prepareMapping($mapping)
-    {
-        $map = require dirname(dirname(__DIR__)) . '/data/mapping/' . $mapping . '.php';
-        foreach ($map as &$term) {
-            if ($term) {
-                $value = explode(':', $term);
-                $term = [$value[0] => $value[1]];
-            } else {
-                $term = [];
-            }
-        }
-        return $map;
-    }
-
-    /**
      * Map Zotero item type to Omeka resource class.
      *
      * @param array $zoteroItem The Zotero item data
      * @param array $omekaItem The Omeka item data
      * @return array
      */
-    public function mapResourceClass(array $zoteroItem, array $omekaItem)
+    protected function mapResourceClass(array $zoteroItem, array $omekaItem)
     {
         if (!isset($zoteroItem['data']['itemType'])) {
             return $omekaItem;
@@ -441,7 +219,7 @@ class Import extends AbstractJob
      * @param array $omekaItem The Omeka item data
      * @return array
      */
-    public function mapValues(array $zoteroItem, array $omekaItem)
+    protected function mapValues(array $zoteroItem, array $omekaItem)
     {
         if (!isset($zoteroItem['data'])) {
             return $omekaItem;
@@ -480,7 +258,7 @@ class Import extends AbstractJob
      * @param array $omekaItem The Omeka item data
      * @return array
      */
-    public function mapNameValues(array $zoteroItem, array $omekaItem)
+    protected function mapNameValues(array $zoteroItem, array $omekaItem)
     {
         if (!isset($zoteroItem['data']['creators'])) {
             return $omekaItem;
@@ -527,7 +305,7 @@ class Import extends AbstractJob
      * @param array $omekaItem The Omeka item data
      * @return array
      */
-    public function mapSubjectValues(array $zoteroItem, array $omekaItem)
+    protected function mapSubjectValues(array $zoteroItem, array $omekaItem)
     {
         if (!isset($zoteroItem['data']['tags'])) {
             return $omekaItem;
@@ -557,7 +335,7 @@ class Import extends AbstractJob
      * @param array $omekaItem The Omeka item data
      * @return string
       */
-    public function mapAttachment($zoteroItem, $omekaItem)
+    protected function mapAttachment($zoteroItem, $omekaItem)
     {
         if ('attachment' === $zoteroItem['data']['itemType']
             && isset($zoteroItem['links']['enclosure'])
@@ -582,34 +360,5 @@ class Import extends AbstractJob
             ];
         }
         return $omekaItem;
-    }
-
-    /**
-     * Get a URL from the Link header.
-     *
-     * @param Response $response
-     * @param string $rel The relationship from the current document. Possible
-     * values are first, prev, next, last, alternate.
-     * @return string|null
-     */
-    public function getLink(Response $response, $rel)
-    {
-        $linkHeader = $response->getHeaders()->get('Link');
-        if (!$linkHeader) {
-            return null;
-        }
-        preg_match_all(
-            '/<([^>]+)>; rel="([^"]+)"/',
-            $linkHeader->getFieldValue(),
-            $matches
-        );
-        if (!$matches) {
-            return null;
-        }
-        $key = array_search($rel, $matches[2]);
-        if (false === $key) {
-            return null;
-        }
-        return $matches[1][$key];
     }
 }
